pyproject.toml
[project]
name = "tucan"
version = "0.1.0"
description = "tucan: Local installable package"
readme = "README.md"
license = { file = "LICENSE" }
authors = [
    { name = "Huub van der Ent", email = "h.t.vanderent@umcutrecht.nl" }
]
requires-python = ">=3.8"

dependencies = [
    "numpy>=1.24",
    "pandas>=2.0",
    "polars>=0.19",
    "torch>=2.0",
    "pyyaml>=6.0"
]

[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[tool.setuptools]
packages = ["tucan"]

[tool.setuptools.package-dir]
"" = "src"

[project.scripts]
tucan = "tucan.cli:entrypoint"
